Final Submission

+++ src/com/service/FileOperation.java
@@ -1,4 +1,4 @@
-package com.main;
+package com.service;
 
 import java.awt.Desktop;
 import java.io.File;
@@ -9,24 +9,28 @@ import java.nio.file.Files;
 import java.nio.file.NoSuchFileException;
 import java.nio.file.NotDirectoryException;
 import java.nio.file.Paths;
-import java.util.Arrays;
 import java.util.LinkedList;
-import java.util.ListIterator;
 
-import javax.swing.text.html.HTMLDocument.Iterator;
-
+/**** User Defined Exception Files****/
 import com.exception.DesktopNotSupportedException;
 import com.exception.DirectoryNotExistException;
-import com.exception.FolderDeletionException;
-import com.notrequired.QuickSort;
-import com.sorting.BinarySearchInLinkedList;
+import com.exception.CannotDeleteFolderException;
+
+/***Sorting and Searching files ****/
+import com.sorting.StringBinarySearch;
 import com.sorting.StringQuickSort;
 
 public class FileOperation {
+
 	private String[] fileNameArray;
 	private LinkedList<String> fileNameList;
+	private String folderPath;
+	private File folderName ;
 
 
+	/**
+	 * setFileNameList Method is used to store the list of files exist in the mentioned folder .
+	 */
 	public void setFileNameList(String[] fileNameArray) {
 		fileNameList = new LinkedList<>();
 		for(String fileName:fileNameArray) {
@@ -34,10 +38,9 @@ public class FileOperation {
 		}		
 	}
 
-
-	private String folderPath;
-	private File folderName ;
-
+	/**
+	 * setFolderPath Method is used to check whether the given folder Path exists or not.
+	 * */
 	public void setFolderPath(String folderPath) throws NotDirectoryException, DirectoryNotExistException {
 		folderName = new File(folderPath);
 		if(folderName.exists()) {
@@ -54,18 +57,18 @@ public class FileOperation {
 	}
 	/**
 	 * createFile Method is used to create the files in the mentioned folder.
+	 * If created the file, then add the file Name to the linked list (fileNameList).
 	 * If the file already exist throws FileAlreadyExistsException.
 	 * @throws FileAlreadyExistsException 
 	 * @throws IOException 
-	 * @throws DirectoryNotExistException 
 	 *
 	 */
-	void createFile(String fileName) throws  FileAlreadyExistsException, IOException {
+	public void createFile(String fileName) throws  FileAlreadyExistsException, IOException {
 		File create= new File(folderPath+"\\"+fileName);
 
 		if(create.createNewFile()) {
 			fileNameList.add(create.getName());
-			sortByAscending(); //Sort the list after adding New File
+			sortByAscending();  //Sort the list after adding New File
 			System.out.println("File Created Successfully");
 		}else {
 			throw new FileAlreadyExistsException("File already exists.Kindly Enter another Name");
@@ -75,25 +78,27 @@ public class FileOperation {
 
 	/**
 	 * deleteFile Method is used to delete the files exist in the mentioned folder.
-	 * @throws DirectoryNotExistException 
+	 * If deleted the file, then delete the file Name from the linked list (fileNameList).
 	 * @throws IOException 
 	 * @throws NoSuchFileException 
-	 * @throws FolderDeletionException 
-	 * @throws NotDirectoryException 
+	 * @throws CannotDeleteFolderException 
 	 *
 	 */
-	void deleteFile(String fileNameToBeDeleted) throws  NoSuchFileException, IOException, FolderDeletionException {
+	public void deleteFile(String fileNameToBeDeleted) throws  NoSuchFileException, IOException, CannotDeleteFolderException {
 		File fileToDelete = new File(folderPath+"\\"+fileNameToBeDeleted);
-		if(fileToDelete.isFile()) {
-			if( Files.deleteIfExists(Paths.get(folderPath+"\\"+fileNameToBeDeleted))) {
-				fileNameList.remove(fileNameToBeDeleted);
+		if(fileToDelete.exists()) {
+			if(fileToDelete.isFile()) {
+				if( Files.deleteIfExists(Paths.get(folderPath+"\\"+fileNameToBeDeleted))) {
+					fileNameList.remove(fileNameToBeDeleted);
+				}else {
+					throw new NoSuchFileException("File does not Exist");
+				}          
 			}else {
-				throw new NoSuchFileException("File does not Exist");
-			}          
+				throw new CannotDeleteFolderException("Can't Delete Folder. ");
+			}
 		}else {
-			throw new FolderDeletionException("Can't Delete Folder. ");
-		}
-
+			throw new NoSuchFileException("File does not Exist");
+		}  
 		System.out.println("Deletion successful.");
 	}
 
@@ -104,30 +109,21 @@ public class FileOperation {
 	 * @throws DirectoryNotExistException 
 	 *
 	 */
-	LinkedList<String> getFileList(){		
-
+	public LinkedList<String> getFileList(){		
 		fileNameArray = folderName.list();	
 		setFileNameList(fileNameArray);
 		LinkedList<String> fileNameLinkedList = fileNameList;
 		return fileNameLinkedList;		
 	}
+
 	/**
 	 * display method is used to display the files available in the given folder.
 	 * 
 	 *
 	 */
-	void display() {
-
-		/*ListIterator<String> iterateList = (ListIterator<String>) fileNameList.iterator();
-		if(fileNameList.size() == 0) {
-			System.out.println("\n \t Folder is Empty !!!!!!!!!");
-		}else {
-			while(iterateList.hasNext()) {
-				System.out.println((iterateList.nextIndex()+1)+"--> "+iterateList.next());
-
-			}
-		}*/
+	public void display() {
 		int index = 1;
+
 		if(fileNameArray.length == 0) {
 			System.out.println("\n \t Folder is Empty !!!!!!!!!");
 		}else {
@@ -139,16 +135,16 @@ public class FileOperation {
 	}
 
 	/**
-	 * searchFile Method is used to search the files in the mentioned folder using
-	 * binarySearch method.
-	 *
+	 * searchFile Method is used to search the given file in the mentioned folder using
+	 * binarySearch method.If file exist in that folder return true and if not then returns
+	 * false.
 	 */
-	boolean searchFile(String fileName) {
-
-		boolean fileExist = BinarySearchInLinkedList.binarySearch(fileNameList, fileName);
+	public boolean searchFile(String fileName) {
+		fileNameArray = getFilesNameArray(fileNameList);
+		boolean fileExist = StringBinarySearch.binarySearch(fileNameArray, fileName);
 
 		if(fileExist) {
-			System.out.println("File "+fileName +" exists. File-path : "+folderPath+"\\"+fileName);
+			System.out.println("File "+fileName +" exists.  File-path : "+folderPath+"\\"+fileName);
 		}else {
 			System.out.println("File "+fileName +" does not exist. \n");
 		}
@@ -156,7 +152,12 @@ public class FileOperation {
 		return fileExist;
 	}
 
-	void openFileInDesktop(String fileName) throws IOException, DesktopNotSupportedException {
+	/**
+	 * openFileInDesktop Method is used to open the given file in the Desktop.
+	 * @throws IOException 
+	 * @throws DesktopNotSupportedException 
+	 */
+	public void openFileInDesktop(String fileName) throws IOException, DesktopNotSupportedException  {
 		File fileToOpen = new File(folderPath+"\\"+fileName);
 
 		if(Desktop.isDesktopSupported()) {
@@ -171,27 +172,40 @@ public class FileOperation {
 		}
 
 	}
-
+	/**
+	 * sortByAscending Method is used to ascend the list of files available in the mentioned folder using
+	 * quickSort method.If the folder is empty then "Folder is Empty". 
+	 * Otherwise return the files list in A-Z Order.sortByDescending Method takes file List as a array.
+	 * As array sorting takes time when compared to sorting a LinkedList.
+	 */
 	public void sortByAscending() {
 		fileNameArray = getFilesNameArray(fileNameList);
 		int listLength = fileNameArray.length-1;
-		System.out.println("List Length"+listLength);
-		if(listLength == -1) {
-			display();
-		}else
+		if(listLength != -1) {
 			StringQuickSort.quickSort(fileNameArray,0, listLength, "Ascending");
 
+		}
 	}
 
+	/**
+	 * sortByDescending Method is used to list files available in the mentioned folder in descending order using
+	 * quickSort method.If the folder is empty then "Folder is Empty". 
+	 * Otherwise return the files list in Z-A Order.  sortByDescending Method takes file List as a array.
+	 * As array sorting takes time when compared to sorting a LinkedList.
+	 */
 	public void sortByDescending() {
 		fileNameArray = getFilesNameArray(fileNameList);
 		int listLength = fileNameArray.length-1;
-		if(listLength == -1) {
-			display();
-		}else
+		if(listLength != -1) {
 			StringQuickSort.quickSort(fileNameArray,0, listLength, "Descending");
+		}
 
 	}
+
+	/**
+	 * getFilesNameArray Method is used to store the updated linked list(fileNameList) 
+	 * fileNames in the array(fileNameArray) before sorting and searching the files.
+	 */
 	public String[] getFilesNameArray(LinkedList<String> fileNameList) {
 		int index = 0;
 		String[]	fileNameArray = new String[fileNameList.size()];
@@ -201,47 +215,4 @@ public class FileOperation {
 		}
 		return fileNameArray;
 	}
-
-	// --------------------------- Not Used  Methods------------------------------//
-
-	/*public void setFilesNameList(String[] fileNameArray) {
-		this.fileNameArray = fileNameArray;
-	}
-
-	private void deleteFileNameFromList(String fileNameToBeDeleted) {
-
-		String[] updatedFilesNameList = new String[fileNameArray.length-1] ;
-		boolean foundFile = false;
-		int indexToBeDeleted = Arrays.binarySearch(fileNameArray, fileNameToBeDeleted);
-
-		if(indexToBeDeleted == fileNameArray.length-1) {
-			for(int index =0;index < fileNameArray.length-1 ;index++) {
-				updatedFilesNameList[index]=fileNameArray[index];
-			}
-		}else {
-			for(int index =0;index < fileNameArray.length-1 ;index++) {
-				if(foundFile || (fileNameArray[index].equals(fileNameToBeDeleted)))  {
-					updatedFilesNameList[index] = fileNameArray[index+1];
-					foundFile = true;
-				}else if(!foundFile){
-					updatedFilesNameList[index]=fileNameArray[index];
-				}
-
-			}
-		}
-
-		setFilesNameList(updatedFilesNameList);
-
-	}
-	private void updateFilesNameList(String newFileName) {
-		String[] updatedfilesNameList =  new String[fileNameArray.length+1];
-		for(int index =0 ;index < fileNameArray.length;index++) {
-			updatedfilesNameList[index] = fileNameArray[index];
-		}
-		updatedfilesNameList[fileNameArray.length] = newFileName;
-		setFilesNameList(updatedfilesNameList);
-
-	}
-
-	 */
 }
